TextUi.java: Extracted formatting responsibility.
New class Formatter has been created to handle the responsibility of formatting the text which TextUi extends from.
